Batch process final marking of found last names
Scanning for last names only is very performant. But when we need to do the
more complex search (via MarkJS) for all permutations of a person's name, it
can get very processor intensive, especially when many names are found within
a single, complex node (such as the document.body).

This chunks the number of last names to be marked, per node, into a queue. The
queue is polled with a setInterval, and if it contains items, they're
processed via the large regex of all permutations of the chunk of last names.

Good for testing: http://www.house.gov/representatives/