Set processing state on extension icon
The only way to animate the icon is by changing it. So this first sets up some
state management for processing (with a debounce function on turning the state
off; so that it's actually visible for half a second). When processing is true,
the background script runs an interval that swaps the icon rapidly to create
a strobing effect.