Fix issue with unicode characters
We're using indexOf in markPermutations to quickly check if each congressData
item's last name was found. Before, we were storing the last name found in the
DOM. So in the case of "Diaz", it would match "Díaz" based upon our prepName()
substitutions, but store "Diaz" as a found last name. But then, while iterating
through the congressData, the item with a last name of "Díaz" would not match
the found string from the DOM.

This goes through found last names and instead stores the original last name
found in the congressData, so we can still do the really fast indexOf lookup
in markPermutations.